python/server/requirements.txt
aicspylibczi==3.1.2
csbdeep==0.7.4
czifile==2019.7.2
Flask==3.0.2
Flask_Cors==4.0.0
imageio==2.33.1
magicgui==0.8.2
matplotlib==3.8.3
napari==0.4.19
numpy==1.26.4
open3d==0.18.0
opencv_contrib_python_headless==4.9.0.80
opencv_python_headless==4.9.0.80
openpyxl==3.1.2
packaging==24.0
pandas==1.5.3
Pillow==9.5.0
Pillow==10.3.0
psygnal==0.11.0
pytest==8.1.1
QtPy==2.4.1
scikit_image==0.19.3
scikit_learn==1.4.1.post1
scipy==1.13.0
setuptools==69.2.0
simpleicp==2.0.14
SimpleITK==2.3.1
SimpleITK==2.3.1
tifffile==2024.1.30
triangle==20230923
valis==0.1
napari-czifile2
